Criando a pagina de formulario.

+++ OrganizerList/app/src/main/java/br/com/organizerlist/model/Product.kt
@@ -3,9 +3,9 @@ package br.com.organizerlist.model
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import br.com.organizerlist.R
-import java.math.BigDecimal
+import java.math.BigInteger
 
 data class Product(
     val name: String,
-    val qty: Int
+    val qty: BigInteger
 )
